failed child node now can be started and in error state
